Initial commit
Quantitative results are displayed in results.txt; files with 'seg' in their name have had the words separated (because in Chinese, there are no spaces to separate words, so I used the Jieba package to separate the words).

Interpreting the results: higher numbers are indicative of superior translation quality (or, if you know Chinese, you can check the quality yourself). The results are percentages (more specifically, they are BLEU scores).

Python files are the code used to conduct this study. There is another version of olympics.py which uses BeautifulSoup instead of manually controlling the mouse and keyboard, but it has insufficient documentation so I will upload it at a later date. Stay tuned.